+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: :warning: Unmarshalling of the rule files is strict again, this behavior was unintentionally brought when adding support for yaml comments.
 - Added: support for alert field `keep_firing_for`
 - Added: support for the `query_offset` field in the rule group
+- Added: new validator `expressionIsValidPromQL` to check if the expression is a valid PromQL query
 
 ## [2.14.1]
 - Fixed: error message in the `hasSourceTenantsForMetrics` validator